requirements.txt
asgiref==3.11.0
Django==6.0
psycopg==3.3.1
sqlparse==0.5.4
typing_extensions==4.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0